+++ epictrack-api/src/api/schemas/response/work_response.py
@@ -151,6 +151,8 @@ class WorkPhaseTemplateAvailableResponse(Schema):
     task_added = fields.Boolean(
         metadata={"description": "Is task added already"}, required=True
     )
+    
+    work_phase = fields.Nested(WorkPhaseResponseSchema, dump_only=True)
 
 
 class WorkStatusResponseSchema(
